+++ ReleaseNotes.md
@@ -1,5 +1,10 @@
 # Release notes
 
+## 1.2.0
+
+- Added JSChannel as dependency in package.json: https://github.com/yochannah/jschannel.git#f2bcb860911d36ccbc3fabd4a09f7074dd9f289e
+- Wrapped OskariRPC.js with UMD pattern https://github.com/umdjs/umd
+
 ## 1.1.0
 
 - added onReady callback to detect when we have a successful connection